requirements.txt
faker==30.8.2
psycopg2-binary==2.9.9
bcrypt==4.2.0
